remove web module, move login to admin, register endpoints in app, add global error handler
